pyproject.toml
[project]
name = "wlyq-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dearpygui>=2.1.0",
    "glfw>=2.9.0",
    "numpy>=2.3.2",
    "pyopengl>=3.1.10",
    "pyopengl-accelerate>=3.1.10",
]
